Add current work for seed rel
