+++ GameCube/source/rando/seed.cpp
@@ -11,76 +11,15 @@
 #include <cstring>
 
 #include "rando/seed.h"
-#include "cxx.h"
 #include "data/stages.h"
-#include "game_patch/game_patch.h"
 #include "gc_wii/card.h"
 #include "main.h"
 #include "rando/data.h"
-#include "tools.h"
-#include "tp/d_a_shop_item_static.h"
 #include "tp/d_com_inf_game.h"
 #include "tp/d_item.h"
-#include "tp/d_item_data.h"
-#include "tp/d_s_logo.h"
-#include "tp/d_save.h"
-#include "user_patch/user_patch.h"
 
 namespace mod::rando
 {
-    Seed::Seed( int32_t chan, SeedInfo* seedInfo ): m_CardSlot( chan )
-    {
-        m_Header = &seedInfo->header;
-        mod::console << m_Header << "\n";
-        // Loading seed rando-dataX '<seed>'...
-
-        // Store our filename index
-        m_fileIndex = seedInfo->fileIndex;
-
-        mod::console << "Loading seed " << m_fileIndex << ": '" << m_Header->seed << "'...\n";
-
-        // Load the whole gci locally to reduce number of reads (memcard)
-        char fileName[12] = "rando-data\0";
-
-        fileName[10] = static_cast<char>( '0' + m_fileIndex );
-
-        // Allocate the buffer to the back of the heap to prevent fragmentation
-        uint32_t totalSize = m_Header->totalSize;
-        uint8_t* data = new ( -0x20 ) uint8_t[totalSize];
-
-        m_CARDResult = libtp::tools::ReadGCI( m_CardSlot, fileName, totalSize, 0, data, true );
-        if ( m_CARDResult == CARD_RESULT_READY )
-        {
-            // Get the main seed data
-            uint32_t dataSize = m_Header->dataSize;
-            m_GCIData = new uint8_t[dataSize];
-            memcpy( m_GCIData, &data[m_Header->headerSize], dataSize );
-
-            // Get the custom text data
-            this->loadCustomText( data );
-        }
-        delete[] data;
-    }
-
-    Seed::~Seed()
-    {
-        // Make sure to delete tempcheck buffers
-        this->ClearChecks();
-
-        // Reset the custom message data
-        m_TotalMsgEntries = 0;
-        delete[] m_MsgTableInfo;
-
-        // Only work with m_GCIData if the buffer is populated
-        if ( m_GCIData )
-        {
-            this->applyPatches( false );
-
-            // Last clear gcibuffer as other functions before rely on it
-            delete[] m_GCIData;
-        }
-    }
-
     bool Seed::InitSeed( void )
     {
         // (Re)set counters & status
@@ -104,95 +43,6 @@ namespace mod::rando
             mod::console << "FATAL: Couldn't read Seed #" << m_fileIndex << "\n";
             mod::console << "ERROR: " << m_CARDResult << "\n";
             return false;
-        }
-    }
-
-    bool Seed::LoadChecks( const char* stage )
-    {
-        using namespace libtp;
-
-        // Find the index of this stage
-        uint8_t stageIDX;
-        for ( stageIDX = 0; stageIDX < sizeof( data::stage::allStages ) / sizeof( data::stage::allStages[0] ); stageIDX++ )
-        {
-            if ( !strcmp( stage, data::stage::allStages[stageIDX] ) )
-            {
-                break;
-            }
-        }
-
-        // Don't run if this isn't actually a new stage
-        bool result = stageIDX != m_StageIDX;
-        if ( result )
-        {
-            this->ClearChecks();
-
-            this->LoadDZX( stageIDX );
-            this->LoadREL( stageIDX );
-            this->LoadPOE( stageIDX );
-            this->LoadBOSS( stageIDX );
-            this->LoadBugReward();
-            this->LoadSkyCharacter( stageIDX );
-            this->LoadHiddenSkill();
-
-            // Save current stageIDX for next time
-            m_StageIDX = stageIDX;
-        }
-
-        return result;
-    }
-
-    void Seed::ClearChecks()
-    {
-        m_numLoadedDZXChecks = 0;
-        m_numLoadedRELChecks = 0;
-        m_numLoadedPOEChecks = 0;
-        m_numLoadedBossChecks = 0;
-        m_numBugRewardChecks = 0;
-        m_numSkyBookChecks = 0;
-        m_numHiddenSkillChecks = 0;
-
-        delete[] m_DZXChecks;
-        delete[] m_RELChecks;
-        delete[] m_POEChecks;
-        delete[] m_BossChecks;
-        delete[] m_BugRewardChecks;
-        delete[] m_SkyBookChecks;
-        delete[] m_HiddenSkillChecks;
-    }
-
-    void Seed::applyPatches( bool set )
-    {
-        using namespace libtp;
-
-        uint32_t num_bytes = m_Header->patchInfo.numEntries;
-        uint32_t gci_offset = m_Header->patchInfo.dataOffset;
-
-        // Don't bother to patch anything if there's nothing to patch
-        if ( num_bytes > 0 )
-        {
-            // Set the pointer as offset into our buffer
-            uint8_t* patch_config = &m_GCIData[gci_offset];
-
-            for ( uint32_t i = 0; i < num_bytes; i++ )
-            {
-                uint8_t byte = patch_config[i];
-
-                for ( uint32_t b = 0; b < 8; b++ )
-                {
-                    if ( ( byte << b ) & 0x80 )
-                    {
-                        // run the patch function for this bit index
-                        uint32_t index = i * 8 + b;
-
-                        if ( index < sizeof( user_patch::patches ) / sizeof( user_patch::patches[0] ) )
-                        {
-                            user_patch::patches[index]( mod::randomizer, set );
-                            m_PatchesApplied++;
-                        }
-                    }
-                }
-            }
         }
     }
 
@@ -289,6 +139,60 @@ namespace mod::rando
                 libtp::tp::d_item::execItemGet( startingItems[i] );
             }
         }
+    }
+
+    bool Seed::LoadChecks( const char* stage )
+    {
+        using namespace libtp;
+
+        // Find the index of this stage
+        uint8_t stageIDX;
+        for ( stageIDX = 0; stageIDX < sizeof( data::stage::allStages ) / sizeof( data::stage::allStages[0] ); stageIDX++ )
+        {
+            if ( !strcmp( stage, data::stage::allStages[stageIDX] ) )
+            {
+                break;
+            }
+        }
+
+        // Don't run if this isn't actually a new stage
+        bool result = stageIDX != m_StageIDX;
+        if ( result )
+        {
+            this->ClearChecks();
+
+            this->LoadDZX( stageIDX );
+            this->LoadREL( stageIDX );
+            this->LoadPOE( stageIDX );
+            this->LoadBOSS( stageIDX );
+            this->LoadBugReward();
+            this->LoadSkyCharacter( stageIDX );
+            this->LoadHiddenSkill();
+
+            // Save current stageIDX for next time
+            m_StageIDX = stageIDX;
+        }
+
+        return result;
+    }
+
+    void Seed::ClearChecks()
+    {
+        m_numLoadedDZXChecks = 0;
+        m_numLoadedRELChecks = 0;
+        m_numLoadedPOEChecks = 0;
+        m_numLoadedBossChecks = 0;
+        m_numBugRewardChecks = 0;
+        m_numSkyBookChecks = 0;
+        m_numHiddenSkillChecks = 0;
+
+        delete[] m_DZXChecks;
+        delete[] m_RELChecks;
+        delete[] m_POEChecks;
+        delete[] m_BossChecks;
+        delete[] m_BugRewardChecks;
+        delete[] m_SkyBookChecks;
+        delete[] m_HiddenSkillChecks;
     }
 
     void Seed::LoadDZX( uint8_t stageIDX )
@@ -395,47 +299,6 @@ namespace mod::rando
                 // Store the i'th POE check into the j'th Loaded POEcheck that's relevant to our current stage
                 memcpy( &m_POEChecks[j], &allPOE[i], sizeof( POECheck ) );
                 j++;
-            }
-        }
-    }
-
-    void Seed::loadShopModels()
-    {
-        using namespace libtp::tp;
-        if ( m_GCIData &&
-             m_Header )     // A fail-safe as the randomizer will crash if it tries to read a seed and there isn't one.
-        {
-            uint32_t num_shopItems = m_Header->shopItemCheckInfo.numEntries;
-            uint32_t gci_offset = m_Header->shopItemCheckInfo.dataOffset;
-
-            // Set the pointer as offset into our buffer
-            shopCheck* allSHOP = reinterpret_cast<shopCheck*>( &m_GCIData[gci_offset] );
-
-            d_item_data::ItemResource* itemResourcePtr = &d_item_data::item_resource[0];
-            for ( uint32_t i = 0; i < num_shopItems; i++ )
-            {
-                if ( allSHOP[i].replacementItemID == libtp::data::items::Foolish_Item )
-                {
-                    game_patch::_02_modifyFoolishShopModel( allSHOP[i].shopItemID );
-                }
-                else
-                {
-                    d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].arcName =
-                        itemResourcePtr[allSHOP[i].replacementItemID].arcName;
-                    d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].modelResIdx =
-                        itemResourcePtr[allSHOP[i].replacementItemID].modelResIdx;
-                    d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].wBckResIdx =
-                        itemResourcePtr[allSHOP[i].replacementItemID].bckResIdx;
-                    d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].wBrkResIdx =
-                        itemResourcePtr[allSHOP[i].replacementItemID].brkResIdx;
-                    d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].wBtpResIdx =
-                        itemResourcePtr[allSHOP[i].replacementItemID].btpResIdx;
-                    d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].tevFrm =
-                        itemResourcePtr[allSHOP[i].replacementItemID].tevFrm;
-                }
-                d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].btpFrm = 0xFF;
-                d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].posY = 15.0f;
-                d_a_shop_item_static::shopItemData[allSHOP[i].shopItemID].mFlags = 0xFFFFFFFF;
             }
         }
     }
@@ -621,74 +484,6 @@ namespace mod::rando
                 j++;
             }
         }
-    }
-
-    bool Seed::loadCustomText( uint8_t* data )
-    {
-        // Get the custom message header
-        CustomMessageHeaderInfo* customMessageHeader =
-            reinterpret_cast<CustomMessageHeaderInfo*>( &data[m_Header->customTextHeaderOffset] );
-
-        // Keep track of the index for the language that is about to be selected
-        uint32_t languageIndex = 0;
-
-        // Get the text for the current language
-#ifdef TP_EU
-        libtp::tp::d_s_logo::Languages lang = libtp::tp::d_s_logo::getPalLanguage2( nullptr );
-        if ( ( lang < libtp::tp::d_s_logo::Languages::uk ) || ( lang > libtp::tp::d_s_logo::Languages::it ) )
-        {
-            // The language is invalid/unsupported, so the game defaults to English
-            lang = libtp::tp::d_s_logo::Languages::uk;
-        }
-        uint32_t language = static_cast<uint32_t>( lang );
-
-        // Get a pointer to the language to use
-        uint32_t totalLanguages = customMessageHeader->totalLanguages;
-        CustomMessageEntryInfo* customMessageInfo = nullptr;
-
-        for ( uint32_t i = 0; i < totalLanguages; i++ )
-        {
-            CustomMessageEntryInfo* entry = &customMessageHeader->entry[i];
-            if ( entry->language == language )
-            {
-                languageIndex = i;
-                customMessageInfo = entry;
-                break;
-            }
-        }
-
-        // If the language wasn't found, then default to English, which should always be the first language included
-        if ( !customMessageInfo )
-        {
-            customMessageInfo = &customMessageHeader->entry[0];
-        }
-#else
-        // US/JP should only have one language included
-        CustomMessageEntryInfo* customMessageInfo = &customMessageHeader->entry[0];
-#endif
-
-        // Allocate memory for the ids, message offsets, and messages
-        uint32_t totalEntries = customMessageInfo->totalEntries;
-        m_TotalMsgEntries = totalEntries;
-
-        uint32_t msgIdTableSize = totalEntries * sizeof( uint16_t );
-        uint32_t msgOffsetTableSize = totalEntries * sizeof( uint32_t );
-
-        // Round msgIdTableSize up to the size of the offsets to make sure the offsets are properly aligned
-        msgIdTableSize = ( msgIdTableSize + sizeof( uint32_t ) - 1 ) & ~( sizeof( uint32_t ) - 1 );
-
-        uint32_t msgTableInfoSize = msgIdTableSize + msgOffsetTableSize + customMessageInfo->msgTableSize;
-        m_MsgTableInfo = new uint8_t[msgTableInfoSize];
-
-        // When calculating the offset the the message table information, we are assuming that the message header is
-        // followed by the entry information for all of the languages in the seed data.
-
-        uint32_t offset = m_Header->customTextHeaderOffset + customMessageInfo->msgIdTableOffset +
-                          ( sizeof( CustomMessageEntryInfo ) * languageIndex ) + sizeof( CustomMessageHeaderInfo );
-
-        // Copy the data to the pointers
-        memcpy( m_MsgTableInfo, &data[offset], msgTableInfoSize );
-        return true;
     }
 
 }     // namespace mod::rando
