+++ GameCube/source/user_patch/01_randombgm.cpp
@@ -149,7 +149,7 @@ namespace mod::user_patch
 
     }     // namespace bgm::bgmrando
 
-    void patchRandomBgm( rando::Randomizer* randomizer, bool set )
+    KEEP_FUNC void patchRandomBgm( rando::Randomizer* randomizer, bool set )
     {
         uint64_t seed = randomizer->getSeed();
         mod::console << "[2] EnemyBgmDisabled [" << ( set ? "x" : " " ) << "]\n";
