+++ GameCube/subrel/seed/source/user_patch/00_wallet.cpp
