+++ GameCube/subrel/seed/source/user_patch/05_newFileFunctions.cpp
