+++ GameCube/subrel/seed/source/user_patch/user_patch.cpp
